[evcc] Move API into package & remove description from httpRequest
