+++ bundles/org.openhab.binding.evcc/src/main/java/org/openhab/binding/evcc/internal/api/EvccAPI.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.evcc.internal;
+package org.openhab.binding.evcc.internal.api;
 
 import static org.openhab.binding.evcc.internal.EvccBindingConstants.EVCC_REST_API;
 import static org.openhab.binding.evcc.internal.EvccBindingConstants.LONG_CONNECTION_TIMEOUT_MILLISEC;
@@ -21,8 +21,8 @@ import java.time.format.DateTimeFormatter;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.evcc.internal.dto.Result;
-import org.openhab.binding.evcc.internal.dto.Status;
+import org.openhab.binding.evcc.internal.api.dto.Result;
+import org.openhab.binding.evcc.internal.api.dto.Status;
 import org.openhab.core.io.net.http.HttpUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -48,15 +48,14 @@ public class EvccAPI {
     /**
      * Make a HTTP request.
      * 
-     * @param description request description for logger
      * @param url full request URL
      * @param method reguest method, e.g. GET, POST
      * @return the response body
      * @throws IOException if HTTP request failed
      */
-    private @Nullable String httpRequest(@Nullable String description, String url, String method) throws IOException {
+    private @Nullable String httpRequest(String url, String method) throws IOException {
         String response = HttpUtil.executeUrl(method, url, LONG_CONNECTION_TIMEOUT_MILLISEC);
-        logger.trace("{} - {}, {} - {}", description, url, method, response);
+        logger.trace("{} {} - {}", method, url, response);
         return response;
     }
 
@@ -67,12 +66,12 @@ public class EvccAPI {
      * Get the status from evcc.
      * 
      * @param host hostname of IP address of the evcc instance
-     * @return {@link org.openhab.binding.evcc.internal.dto.Result} object or null
+     * @return {@link org.openhab.binding.evcc.internal.api.dto.Result} object or null
      * @throws IOException if HTTP request failed
      * @throws JsonSyntaxException if response JSON read failed
      */
     public @Nullable Result getResult() throws IOException, JsonSyntaxException {
-        final String response = httpRequest("Status", this.host + EVCC_REST_API + "state", "GET");
+        final String response = httpRequest(this.host + EVCC_REST_API + "state", "GET");
         Status status = gson.fromJson(response, Status.class);
         if (status == null) {
             return null;
@@ -82,44 +81,36 @@ public class EvccAPI {
 
     // Loadpoint specific API calls.
     public @Nullable String setMode(int loadpoint, String mode) throws IOException {
-        return httpRequest("Set mode of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/mode/" + mode, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/mode/" + mode, "POST");
     }
 
     public @Nullable String setMinSoC(int loadpoint, int minSoC) throws IOException {
-        return httpRequest("Set minSoC of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/minsoc/" + minSoC, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/minsoc/" + minSoC, "POST");
     }
 
     public @Nullable String setTargetSoC(int loadpoint, int targetSoC) throws IOException {
-        return httpRequest("Set targetSoC of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/targetsoc/" + targetSoC, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/targetsoc/" + targetSoC, "POST");
     }
 
     public @Nullable String setPhases(int loadpoint, int phases) throws IOException {
-        return httpRequest("Set phases of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/phases/" + phases, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/phases/" + phases, "POST");
     }
 
     public @Nullable String setMinCurrent(int loadpoint, int minCurrent) throws IOException {
-        return httpRequest("Set minCurrent of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/mincurrent/" + minCurrent, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/mincurrent/" + minCurrent, "POST");
     }
 
     public @Nullable String setMaxCurrent(int loadpoint, int maxCurrent) throws IOException {
-        return httpRequest("Set maxCurrent of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/maxcurrent/" + maxCurrent, "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/maxcurrent/" + maxCurrent, "POST");
     }
 
     public @Nullable String setTargetCharge(int loadpoint, int targetSoC, ZonedDateTime targetTime) throws IOException {
         DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-        return httpRequest("Set targetTime of loadpoint " + loadpoint, this.host + EVCC_REST_API + "loadpoints/"
-                + loadpoint + "/targetcharge/" + targetSoC + "/" + targetTime.toLocalDateTime().format(formatter),
-                "POST");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/targetcharge/" + targetSoC + "/"
+                + targetTime.toLocalDateTime().format(formatter), "POST");
     }
 
     public @Nullable String unsetTargetCharge(int loadpoint) throws IOException {
-        return httpRequest("Unset targetTime of loadpoint " + loadpoint,
-                this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/targetcharge", "DELETE");
+        return httpRequest(this.host + EVCC_REST_API + "loadpoints/" + loadpoint + "/targetcharge", "DELETE");
     }
 }
